Implement Exercise: Packaging Python Code
