+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup
+import setuptools
+
+if __name__ == "__main__":
+    setup(url='https://github.com/ProbstenHias/diffusion2d',
+          author='Matthias Weilinger',
+          author_email='st156232@stud.uni-stuttgart.de',
+          )
